test: replaced test_utils with test_utils_new
