+++ test/integration/client/test_ibkr_ws_client_i.py
@@ -10,7 +10,7 @@ from ibind import Result
 from ibind.client.ibkr_client import IbkrClient
 from ibind.client.ibkr_ws_client import IbkrWsClient, IbkrSubscriptionProcessor, IbkrWsKey
 from integration.base.websocketapp_mock import create_wsa_mock, init_wsa_mock
-from test_utils_new import capture_logs
+from test_utils import capture_logs
 
 _URL_WS = 'wss://localhost:5000/v1/api/ws'
 _URL_REST = 'https://localhost:5000'
